requirements.txt
pyautogui
opencv-python 
Pillow
numpy
opencv-python
tensorflow
imageai --upgrade

